// library/src/lib/lib/types.ts
export type TriageStatus =
  | "Awaiting Disposition"
  | "In Progress"
  | "Pending Clear Scan"
  | "Resolved";

export type SeverityRisk =
  | "Priority 1"
  | "Priority 2"
  | "Priority 3"
  | "Priority 4";

export type SourceStatus = "Open" | "Closed";

export type Disposition =
  | "Fix"
  | "Defer"
  | "Mitigate"
  | "Accept Risk"
  | "False Positive"
  | "";

export type OperatingEnvironment = "In Production" | "Pre-Prod" | "Contingency";